+++ Innowise.Music/ViewModel/LoginPageViewModel.cs
@@ -1,10 +1,18 @@
-﻿using CommunityToolkit.Mvvm.ComponentModel;
+using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
+using Innowise.Music.Services;
 
 namespace Innowise.Music.ViewModel;
 
 public partial class LoginPageViewModel : ObservableObject
 {
+    private readonly INavigationService _navigationService;
+
+    public LoginPageViewModel(INavigationService navigationService)
+    {
+        _navigationService = navigationService;
+    }
+
     [RelayCommand]
     private async Task Login()
     {
@@ -14,6 +22,6 @@ public partial class LoginPageViewModel : ObservableObject
     [RelayCommand]
     private async Task SignUp()
     {
-        await Shell.Current.GoToAsync(nameof(View.SignUpPage));
+        await _navigationService.NavigateToAsync(nameof(View.SignUpPage));
     }
 }